requirements.txt
numpy==1.24.3
pandas==1.5.3
Flask==2.2.5
nbformat
plotly.express
ipykernel
gunicorn==23.0.0
requests
#flask